+++ common/src/main/java/TCOTS/mixin/WorldRendererMixin.java
@@ -1,6 +1,8 @@
 package TCOTS.mixin;
 
 import TCOTS.registry.TCOTS_Effects;
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.sugar.Local;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
@@ -23,7 +25,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Environment(value= EnvType.CLIENT)
@@ -60,13 +61,9 @@ public abstract class WorldRendererMixin {
         }
     }
 
-    @Redirect(method= "renderLevel", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;shouldEntityAppearGlowing(Lnet/minecraft/world/entity/Entity;)Z"))
-    private boolean injectCatEffectOutline(Minecraft instance, Entity entity){
-        if(this.shouldShowEntityOutlines() && this.canHaveCatEffect() && checkEntity(entity) ){
-         return true;
-        }
-
-        return this.minecraft.shouldEntityAppearGlowing(entity);
+    @ModifyExpressionValue(method= "renderLevel", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/Minecraft;shouldEntityAppearGlowing(Lnet/minecraft/world/entity/Entity;)Z"))
+    private boolean injectCatEffectOutline(boolean original, @Local Entity entity){
+        return original ||(this.shouldShowEntityOutlines() && this.canHaveCatEffect() && checkEntity(entity));
     }
 
     @Unique